+++ docs/reference/changelog.rst
@@ -44,6 +44,9 @@ Models
   how these fields should behave in their apps.
 - ``any`` and ``*`` are now reserved names for platforms.
 - Platform names may not contain '/' characters.
+- Ubuntu Questing is now a 'stable' base, and Resolute is added as the new 'devel' one.
+- Add a classmethod on the :py:class:`~craft_application.models.Project` class to let
+  applications fine-tune which bases are considered 'devel'.
 
 Fetch Service
 ~~~~~~~~~~~~~